+++ src/components/sidebar/RulesCtrl.tsx
@@ -196,7 +196,7 @@ export default defineComponent({
         <TextInput
           class={isLargeCtrlsBar.value ? 'w-80' : 'w-32 flex-1'}
           v-model={rulesFilter.value}
-          placeholder={`${t('search')} | ${t('searchMultiple')}`}
+          placeholder={`${t('search')} 域名 / IP / 关键字`}
           clearable={true}
         />
       )